feat(): linked todos to user accounts

+++ convex/functions.ts
@@ -1,9 +1,11 @@
 import { mutation, query } from "./_generated/server";
 import { v } from "convex/values";
+import { requireUser } from "./helpers";
 
 export const listTodos = query({
     handler: async (ctx) => {
-        return await ctx.db.query("todos").collect();
+        const user = await requireUser(ctx);
+        return await ctx.db.query("todos").withIndex("by_user_id", q => q.eq("userId", user.tokenIdentifier)).collect();
     }
 });
 
@@ -13,10 +15,12 @@ export const createTodo = mutation({
         description: v.string()
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
         await ctx.db.insert("todos", {
             title: args.title,
             description: args.description,
-            completed: false
+            completed: false,
+            userId: user.tokenIdentifier,
         });
     },
 });
@@ -27,6 +31,11 @@ export const updateTodo = mutation({
         completed: v.boolean()
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id);
+        if (todo?.userId !== user.tokenIdentifier) {
+            throw new Error("Unauthorized action")
+        }
         await ctx.db.patch(args.id, {
             completed: args.completed
         });
@@ -38,6 +47,11 @@ export const removeTodo = mutation({
         id: v.id("todos")
     },
     handler: async (ctx, args) => {
+        const user = await requireUser(ctx);
+        const todo = await ctx.db.get(args.id);
+        if (todo?.userId !== user.tokenIdentifier) {
+            throw new Error("Unauthorized action")
+        }
         await ctx.db.delete(args.id);
     },
 });
